+++ src/standalone/scripts/green-bean-checks.ts
@@ -1,6 +1,6 @@
 import { NestFactory } from '@nestjs/core';
-import { AppModule } from '../app.module';
-import { GreenBeanService } from '../services/green-bean.service';
+import { AppModule } from '../../app.module';
+import { GreenBeanService } from '../../services/green-bean.service';
 
 // Initialize the db with canClaim statuses for all Azukis
 async function bootstrap() {
